Inline policy::concrete module
Now we have done re-names and the concrete policy is the 'first class'
type in `policy` inline the logic from the submodle to `policy/mod.rs`.
